move CubedSphereDivider and tests to fv3fit.reservoir
