+++ external/fv3fit/fv3fit/reservoir/__init__.py
@@ -9,3 +9,4 @@ from .model import ReservoirComputingModel, HybridReservoirComputingModel
 from .domain import RankDivider
 from .transformers.autoencoder import Autoencoder
 from .transformers.sk_transformer import SkTransformer
+from .cubed_sphere import CubedSphereDivider
